healthcare-api/pyproject.toml
[project]
name = "healthcare-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno>=2.4.7",
    "fastapi==0.109.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic==2.5.3",
    "pymongo==4.6.1",
    "python-dotenv==1.0.0",
    "python-jose>=3.5.0",
    "python-multipart==0.0.6",
    "uvicorn==0.27.0",
]
